fix: bug hunt passe 2 — 20 correctifs, dont le filet de test lui-même
4 agents sur des angles jamais balayés (CPU/Moira, outils de débogage GUI,
périphériques secondaires, OUTILLAGE de test) + traitement du reliquat vérifié
de la passe 1. Étalons pixel verts avant/après.

LE FILET DE SÉCURITÉ AVAIT DES TROUS (le plus grave de la passe : ces outils
sont la SEULE validation du projet, cf. CLAUDE.md)
- run_selftests.py rendait VERT un émulateur qui SEGFAUTE : code de retour
  jamais lu, et le dump série n'étant écrit qu'en fin de main(), celui du run
  précédent était relu comme neuf. Prouvé avec un faux binaire qui se tue :
  « TOUS OK », exit 0. Désormais dump effacé avant le run, rc vérifié, stderr
  conservé dans tests/out/<id>_run.log.
- « --only <ID inconnu> » exécutait ZÉRO test et rendait « TOUS OK » exit 0.
  run_all.py passant au palier P0 une liste d'IDs CODÉE EN DUR, un simple
  renommage dans le manifeste aurait vidé le garde-fou du hook pre-push sans
  un mot. Les deux runners refusent maintenant un ID inconnu (exit 2), et un
  filtre qui ne sélectionne rien est une erreur.
- Une référence absente était comptée comme RÉUSSITE : cuddly_demos, l'étalon
  bordures de référence du projet, est aujourd'hui un no-op vert. Le
  récapitulatif liste désormais « ⚠ NON COMPARÉS ».

SAVE-STATES (v6 → v7)
- Un état sauvé avec le HD GEMDOS était ACCEPTÉ dans une session sans : _drvbits
  annonçait toujours C: et le vecteur trap #1 pointait dans un port cartouche
  vide ($FFFFFFFF) — premier appel GEMDOS dans le décor. GemdosHd n'ayant aucun
  serialize, l'empreinte le REFUSE désormais, cartouche comprise (empreinte
  CRC32), et `bus.cart` est sérialisée : elle porte l'ancien vecteur GEMDOS,
  donc de l'état muté, pas une image immuable.

STX — pertes de données prouvées
- les WRITE SECTOR faits après un WRITE TRACK disparaissaient au remontage
  (blocs SECT lus avant les TRCK, et recherche dans les secteurs d'origine au
  lieu de la vue réinterprétée) : la sauvegarde d'un jeu qui formate sa piste
  était perdue au lancement suivant. Lecture en deux passes + sectorsView().
- overlays orphelins jamais invalidés → .wd1772 qui grossit sans fin ;
- overlay dimensionné sur le transfert et relu sur sec.sectorSize → lecture
  hors du vector si la face change en cours de transfert.

CRASHS / ROBUSTESSE
- 3 scans de dossiers utilisaient is_regular_file() SANS error_code : un
  symlink dont la cible devient illisible (clé USB débranchée) lance une
  filesystem_error que rien n'attrape → std::terminate. La Cart Library étant
  ouverte par défaut et rescannée à chaque trame, le crash est immédiat.
- Blitter : compteur d'accès bus en int → UB au-delà de 2,1e9 accès (x=y=0
  vaut 65536×65536), et le stall CPU débordait dès 5,4e8, donc perdu.
- Fdc::loadImage ne vérifiait pas gcount() : une lecture courte donnait une
  image complétée de zéros, montée comme valide.
- WASM : shader non compilé attaché quand même, link jamais vérifié → attribs
  à -1 passés à chaque trame, canvas noir sans diagnostic.

Divers : --dump-at jamais atteint et --from-cfg sur un répertoire rendaient 0
(verts silencieux pour un runner d'oracle) ; invariants save-state YM ;
touche K du kiosk envoyée aussi au ST (le bandeau avalait ensuite le clavier) ;
Disk Library comparant le nom de fichier au lieu du chemin (deux dumps
homonymes → le second inaccessible) ; passe CRT du kiosk demandée à la taille
écran alors que le viewport l'agrandit ensuite (moiré du masque).

NON APPLIQUÉ, DOCUMENTÉ SUR PLACE : le double comptage de Machine::liveNow()
pendant le dispatch du saut STOP (1208 dispatches sur 15780, δ ≤ 112 cyc,
mesuré et reproduit). Le correctif — deux lignes à remonter — annule
l'incohérence (1208 → 0) mais CASSE l'étalon oracle nocooper (10828 px) : les
datations empiriques −6/+2/−25 ont été calibrées AVEC ce défaut et l'absorbent.
Corriger la cause impose de refaire la calibration contre l'oracle Hatari.

+++ src/core/Blitter.hpp
@@ -86,7 +86,7 @@ private:
     void start();                            // BUSY écrit à 1 : démarre/reprend le blit
     bool runSlice(int maxBusAccesses);       // ≤ N accès bus (-1 = tout) ; true = terminé
     void finishTransfer();                   // y_count = 0 : BUSY/HOG effacés + IRQ GPIP3
-    void stallCpu(int busAccesses, int arbCycles);   // 4 cyc/accès + arbitration (Moira)
+    void stallCpu(int64_t busAccesses, int arbCycles);   // 4 cyc/accès + arbitration (Moira)
     void pauseTransfer();                    // BUSY effacé pendant un blit : tranche annulée
     // Fenêtre PRE_START de 4 cycles avant chaque prise de bus non-hog (cf. .cpp) :
     // armée dans Bus::blitterWinStart/End, consultée par les callbacks mémoire de
@@ -146,7 +146,10 @@ private:
     bool     haveDst_  = false;              // lecture destination du mot courant déjà faite
     bool     fetchSrc_ = false;              // une source a été lue (màj src_addr en fin de mot)
     uint16_t dstWord_  = 0;                  // mot destination lu (survit à la suspension)
-    int      sliceBus_ = 0;                  // accès bus consommés par la tranche en cours
+    // int64 : un blit HOG x_count=y_count=0 vaut 65536×65536 → ~8,6e9 accès bus, soit
+    // un débordement signé (UB) sur int, et « busAccesses * 4 » débordait dès 5,4e8 —
+    // le stall CPU devenait alors négatif et était silencieusement ignoré.
+    int64_t  sliceBus_ = 0;                  // accès bus consommés par la tranche en cours
     int      cpuBusCnt_ = 0;                 // accès bus CPU consommés (fenêtre COUNT_CPU_BUS)
     bool     busCountError_ = false;         // accès CPU « volé » en PRE_START → tranche de 63
 };
